ComplienceMonitoringSystem-main/backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flower==2.0.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
kombu==5.5.3
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.5
pandas==2.2.3
priority==1.3.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
service-identity==24.2.0
setuptools==80.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
